new_project/requirements.txt
# Flask Web Framework
Flask==2.3.0
flask-cors==4.0.0

# Data Processing
pandas==2.0.0
numpy==1.24.0

# Machine Learning
scikit-learn==1.3.0
tensorflow==2.13.0

# GPU Acceleration (Optional)
cupy-cuda12x==12.0.0  # 如果有NVIDIA GPU

# Visualization
matplotlib==3.7.0
plotly==5.14.0

# Utils
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.0
pytest-cov==4.1.0
